IRL Lighting Rewrite
This breaks the IRL lighting into 3 parts (Controller, Interpreter, Hardware) so that things can be expanded, extended, and added too much more easily for future support of more things.
