+++ Assets/Scenes/PersistentScene.unity
@@ -982,9 +982,11 @@ GameObject:
   m_Component:
   - component: {fileID: 728815264}
   - component: {fileID: 728815265}
-  - component: {fileID: 728815266}
   - component: {fileID: 728815267}
   - component: {fileID: 728815269}
+  - component: {fileID: 728815271}
+  - component: {fileID: 728815273}
+  - component: {fileID: 728815272}
   m_Layer: 0
   m_Name: Integrations
   m_TagString: Untagged
@@ -1019,23 +1021,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 43233795acd9ae84690111f3010c5413, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &728815266
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 728815263}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1d1faa5ee2902ba40bafec9aea861ebd, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  CurrentFogState: 0
-  PreviousFogState: 0
-  CurrentStrobeState: 0
-  PreviousStrobeState: 0
-  LargeVenue: 0
 --- !u!114 &728815267
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1058,6 +1043,43 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 525ce11d5f9f48a48adee4bc7515712a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &728815271
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 728815263}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e549df73151ee6b48a2700b69f0e6be2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  <GameManager>k__BackingField: {fileID: 0}
+--- !u!114 &728815272
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 728815263}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1d1faa5ee2902ba40bafec9aea861ebd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &728815273
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 728815263}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2e6d233a733a6014891836deea6c0bc2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &782876828
