+++ Assets/Script/Integration/MasterLightingController.cs.meta
