+++ Assets/Script/Integration/StageKit/StageKitHardware.cs.meta
