+++ Assets/Script/Integration/StageKit/StageKitInterpreter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d7ca8450042ddbb4fb97489d8d14cdc3
+guid: e549df73151ee6b48a2700b69f0e6be2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
