+++ Assets/Script/Integration/StageKit/StageKitLighting.Primitives.cs
@@ -12,10 +12,10 @@ namespace YARG.Integration.StageKit
     {
         private readonly bool _continuous;
         private int _patternIndex = 0;
-        private readonly (int color, byte data)[] _patternList;
+        private readonly (StageKitLedColor color, byte data)[] _patternList;
         private GameManager _gameManager;
 
-        public BeatPattern((int, byte)[] patternList, float beatsPerCycle, bool continuous = true)
+        public BeatPattern((StageKitLedColor, byte)[] patternList, float beatsPerCycle, bool continuous = true)
         {
             //Brought to you by Hacky Hack and the Hacktones
             _gameManager = Object.FindObjectOfType<GameManager>();
@@ -27,7 +27,7 @@ namespace YARG.Integration.StageKit
 
         public override void OnBeat()
         {
-            StageKitLightingController.Instance.SetLed(_patternList[_patternIndex].color,
+            StageKitInterpreter.Instance.SetLed(_patternList[_patternIndex].color,
                 _patternList[_patternIndex].data);
             _patternIndex++;
 
@@ -36,7 +36,7 @@ namespace YARG.Integration.StageKit
             if (!_continuous && _patternIndex == _patternList.Length)
             {
                 _gameManager.BeatEventHandler.Unsubscribe(OnBeat);
-                StageKitLightingController.Instance.CurrentLightingCue.CuePrimitives.Remove(this);
+                StageKitInterpreter.Instance.CurrentLightingCue.CuePrimitives.Remove(this);
             }
 
             if (_patternIndex >= _patternList.Length)
@@ -50,11 +50,12 @@ namespace YARG.Integration.StageKit
     {
         private readonly ListenTypes _listenType;
         private int _patternIndex;
-        private readonly (int color, byte data)[] _patternList;
+        private readonly (StageKitLedColor color, byte data)[] _patternList;
         private readonly bool _flash;
         private readonly bool _inverse;
 
-        public ListenPattern((int color, byte data)[] patternList, ListenTypes listenType, bool flash = false,
+        public ListenPattern((StageKitLedColor color, byte data)[] patternList, ListenTypes listenType,
+            bool flash = false,
             bool inverse = false)
         {
             _flash = flash;
@@ -63,7 +64,7 @@ namespace YARG.Integration.StageKit
             _inverse = inverse;
 
             if (!_inverse) return;
-            StageKitLightingController.Instance.SetLed(_patternList[_patternIndex].color,
+            StageKitInterpreter.Instance.SetLed(_patternList[_patternIndex].color,
                 _patternList[_patternIndex].data);
             _patternIndex++;
 
@@ -106,16 +107,18 @@ namespace YARG.Integration.StageKit
 
         private void ProcessEvent()
         {
-            //This might be a bug in the official  stagekit code i'm trying to replicate here, but instead of turning off the strobe as soon as cue changes, if the cue listens for something, it only turns off the strobe on the first event of it.
-            StageKitLightingController.Instance.SetStrobeSpeed(StageKitStrobeSpeed.Off);
+            //This might be a bug in the official stage kit code. Instead of turning off the strobe as soon as cue
+            //changes, if the cue listens for something, it only turns off the strobe on the first event of it.
+            //To make that happen, strobe off would have to be here and removed from the master controller as well as
+            //added to the lighting event switch case for all the non-listening cues.
 
             if (_inverse)
             {
-                StageKitLightingController.Instance.SetLed(_patternList[_patternIndex].color, NONE);
+                StageKitInterpreter.Instance.SetLed(_patternList[_patternIndex].color, NONE);
             }
             else
             {
-                StageKitLightingController.Instance.SetLed(_patternList[_patternIndex].color,
+                StageKitInterpreter.Instance.SetLed(_patternList[_patternIndex].color,
                     _patternList[_patternIndex].data);
             }
 
@@ -139,12 +142,12 @@ namespace YARG.Integration.StageKit
             await UniTask.Delay(200);
             if (_inverse)
             {
-                StageKitLightingController.Instance.SetLed(_patternList[_patternIndex].color,
+                StageKitInterpreter.Instance.SetLed(_patternList[_patternIndex].color,
                     _patternList[_patternIndex].data);
             }
             else
             {
-                StageKitLightingController.Instance.SetLed(_patternList[_patternIndex].color, NONE);
+                StageKitInterpreter.Instance.SetLed(_patternList[_patternIndex].color, NONE);
             }
         }
     }
@@ -153,9 +156,9 @@ namespace YARG.Integration.StageKit
     {
         private readonly float _seconds;
         private int _patternIndex;
-        private readonly (int color, byte data)[] _patternList;
+        private readonly (StageKitLedColor color, byte data)[] _patternList;
 
-        public TimedPattern((int, byte)[] patternList, float seconds)
+        public TimedPattern((StageKitLedColor, byte)[] patternList, float seconds)
         {
             //Token only for timed events
             CancellationTokenSource = new CancellationTokenSource();
@@ -168,11 +171,14 @@ namespace YARG.Integration.StageKit
         {
             while (!cancellationToken.IsCancellationRequested)
             {
-                StageKitLightingController.Instance.SetLed(_patternList[_patternIndex].color,
+                StageKitInterpreter.Instance.SetLed(_patternList[_patternIndex].color,
                     _patternList[_patternIndex].data);
+
                 await UniTask.Delay(TimeSpan.FromSeconds(_seconds / _patternList.Length),
                     cancellationToken: cancellationToken);
+
                 _patternIndex++;
+
                 if (_patternIndex >= _patternList.Length)
                 {
                     _patternIndex = 0;
@@ -181,3 +187,11 @@ namespace YARG.Integration.StageKit
         }
     }
 }
+/*
+    "One thing kids like is to be tricked. For instance, I was going to take my little nephew to Disneyland, but instead
+    I drove him to an old burned-out warehouse. “Oh, no,” I said.  “Disneyland burned down.”  He cried and cried, but I
+    think that deep down, he thought it was a pretty good joke.  I started to drive over to the real Disneyland, but it
+    was getting pretty late."
+
+    - Jack Handey
+*/
